#234 moved tests for included views #235
