--- mvvmfx/src/test/resources/de/saxsys/mvvmfx/resourcebundle/IncludedView.fxml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-
-<VBox fx:controller="de.saxsys.mvvmfx.resourcebundle.IncludedView" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" xmlns="http://javafx.com/javafx/8.0.40" xmlns:fx="http://javafx.com/fxml/1">
-   <children>
-      <Label fx:id="label" text="%label" />
-   </children>
-</VBox>
