--- mvvmfx/src/test/resources/de/saxsys/mvvmfx/resourcebundle/RootView.fxml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.scene.layout.*?>
-
-
-<VBox fx:controller="de.saxsys.mvvmfx.resourcebundle.RootView" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" xmlns="http://javafx.com/javafx/8.0.40" xmlns:fx="http://javafx.com/fxml/1">
-	<children>
-		<fx:include fx:id="includedView" source="IncludedView.fxml" resources="sub"/>
-	</children>
-</VBox>
